pyproject.toml
[tool.poetry]
name = "tec-welcome-bot"
version = "0.1.0"
description = "A bot to welcome and onboard people to the TEC."
authors = ["Vyvy-vi <vyom.j@protonmail.com>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = ">=3.8.0,<3.10"
aiohttp = "^3.7.4"
uvloop = "^0.16.0"
hikari = {extras = ["speedups"], version = "^2.0.0-alpha.101"}
hikari-lightbulb = "^1.2.4"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.0"
flake8 = "^3.9.2"
black = "^21.8b0"
isort = "^5.9.3"
taskipy = "^1.8.2"

[tool.poetry.scripts]
start = "bot:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


